% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcEnrichment.R
\name{CalcEnrichment}
\alias{CalcEnrichment}
\title{CalcEnrichment.}
\usage{
CalcEnrichment(x = c(100, 10, 1), sig = 4, robust = 0)
}
\arguments{
\item{x}{Vector of intensities.}

\item{sig}{Round enrichment to this precision.}

\item{robust}{Provide an intensity threshold that at least one ion has to exceed. Otherwise NA is returned.}
}
\value{
The enrichment, defined as the ratio of 13C/total_C.
}
\description{
\code{CalcEnrichment} will calculate from a vector of isotope intensities the enrichment of 13C.
}
\keyword{internal}
