% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CandidateBoxplot.R
\name{CandidateBoxplot}
\alias{CandidateBoxplot}
\title{CandidateBoxplot.}
\usage{
CandidateBoxplot(res = NULL)
}
\arguments{
\item{res}{Candidate evaluation result.}
}
\value{
A list with all important information including deconvoluted spectrum and linear model.
}
\description{
\code{CandidateBoxplot} will plot enrichment of tracer incorporationn candidates.
}
\details{
Plot will be annotated with mz/rt information plus ANOVA(lm()) P-values.
}
\keyword{internal}
