% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorMID_reexport.R
\docType{import}
\name{CorMID}
\alias{CorMID}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{CorMID}{\code{\link[CorMID]{CorMID}}}
}}

