% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat-data.R
\docType{data}
\name{dat}
\alias{dat}
\title{Raw data.}
\description{
This list contains a small time and mass window of raw data from APCI measurements as imported by \code{xcms}.
}
\author{
Jan Lisec \email{jan.lisec@bam.de}
}
\keyword{data}
