% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DetermineElementComposition.R
\name{DetermineElementComposition}
\alias{DetermineElementComposition}
\title{DetermineElementComposition.}
\usage{
DetermineElementComposition(int = NULL, mz = 300, ionization = "APCI")
}
\arguments{
\item{int}{Measured intensities.}

\item{mz}{Mass ob peak.}

\item{ionization}{Currently only APCI is supported.}
}
\value{
The most likely combinations of C and Si atoms for this formula.
}
\description{
\code{DetermineElementComposition} will compute the Mass Distribution Vectors of isotopologues based on measured intensities at [13C]=1\%.
}
\details{
Not yet.
}
\keyword{internal}
