% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportHeuristicsTable.R
\name{ExportHeuristicsTable}
\alias{ExportHeuristicsTable}
\title{ExportHeuristicsTable.}
\usage{
ExportHeuristicsTable(res_list1 = NULL, res_list2 = NULL, out1 = NULL,
  out2 = NULL, p_thr = 0.01, xls_name = "Tables.xlsx")
}
\arguments{
\item{res_list1}{res_list1.}

\item{res_list2}{res_list2.}

\item{out1}{out1.}

\item{out2}{out2.}

\item{p_thr}{p_thr.}

\item{xls_name}{xls_name.}
}
\description{
\code{ExportHeuristicsTable}.
}
\keyword{internal}
