% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupval.R
\name{groupval}
\alias{groupval}
\title{groupval.}
\usage{
groupval(xg = NULL)
}
\arguments{
\item{xg}{xcmsSet object.}
}
\description{
\code{groupval} is a dirty rewrite of the same function available in xcms.
}
\details{
[20190514 JL xcms needs to be put to suggest due to mzR on request by CRAN --> replaced groupval from xcms by quick and dirty own function]
}
\keyword{internal}
