% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiResTEC-data.R
\docType{data}
\name{raw}
\alias{raw}
\alias{sam}
\alias{xcms_cand}
\alias{res_list}
\alias{mz_shift_corrector}
\title{Different example data files to be used in the help section of `HiResTEC` functions.}
\format{
An object of class \code{list} of length 18.

An object of class \code{data.frame} with 18 rows and 8 columns.

An object of class \code{data.frame} with 72 rows and 19 columns.

An object of class \code{list} of length 14.

An object of class \code{list} of length 3.
}
\source{
jan.lisec@bam.de
}
\usage{
data(raw)

data(sam)

data(xcms_cand)

data(res_list)

data(mz_shift_corrector)
}
\description{
`raw` contains a list of `xcmsRawLike` objects, each containing a selected mass range for a small retention time window of 18 samples defined in sam.

`sam` provides a data frame containing the sample definition of 18 samples from a larger experiment.

`xcms_cand` contains a data frame with the analysis result of an `xcmsSet` obtained using \link{EvaluatePairsFromXCMSSet}.

`res_list` is a list containing the evaluations results established based on processing example data with \link{EvaluateCandidateListAgainstRawData}.

`mz_shift_corrector` contains a list defining windows for high res APCI or ESI instrumentation.
}
\keyword{data}
