% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiResTEC-data.R
\docType{data}
\name{mz_shift_corrector}
\alias{mz_shift_corrector}
\title{Predefined mass search windows to be used internally.}
\format{
An object of class \code{list} of length 3.
}
\source{
jan.lisec@bam.de
}
\usage{
data(mz_shift_corrector)
}
\description{
This is a list defining windows for high res APCI or ESI instrumentation..
}
\keyword{data}
