% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RankCandidateList.R
\name{RankCandidateList}
\alias{RankCandidateList}
\title{RankCandidateList.}
\usage{
RankCandidateList(x)
}
\arguments{
\item{x}{Candidate dataframe.}
}
\value{
Reordered candidate data.frame.
}
\description{
\code{RankCandidateList} will reorder a candidate list of mz pairs according to maximum intensity.
}
\details{
If still present, rows containing NA in mz2 column will be filtered additionally.
}
\keyword{internal}
