% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiResTEC-data.R
\docType{data}
\name{raw}
\alias{raw}
\title{Mass Spectrometry measurement data}
\format{
An object of class \code{list} of length 18.
}
\source{
jan.lisec@bam.de
}
\usage{
data(raw)
}
\description{
This contains a list of `xcmsRawLike` objects containing a selected mass range for a small retention time window of 18 samples defined in sam.
}
\keyword{data}
