% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiResTEC-data.R
\docType{data}
\name{res_list}
\alias{res_list}
\title{The main results object of a non-targeted search for tracer incorporation.}
\format{
An object of class \code{list} of length 14.
}
\source{
jan.lisec@bam.de
}
\usage{
data(res_list)
}
\description{
This is a list containing the evaluations results established based on processing example data with \link{EvaluateCandidateListAgainstRawData}.
}
\keyword{data}
