% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiResTEC-data.R
\docType{data}
\name{sam}
\alias{sam}
\title{Sample table}
\format{
An object of class \code{data.frame} with 18 rows and 8 columns.
}
\source{
jan.lisec@bam.de
}
\usage{
data(sam)
}
\description{
This data frame contains the sample definition of 18 samples from a larger experiment.
}
\keyword{data}
