% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiResTEC-data.R
\docType{data}
\name{xcms_cand}
\alias{xcms_cand}
\title{Dataframe with putative candidates}
\format{
An object of class \code{data.frame} with 72 rows and 19 columns.
}
\source{
jan.lisec@bam.de
}
\usage{
data(xcms_cand)
}
\description{
This data frame contains the analysis result of an xcmsSet which can not be provided via CRAN anymore using \link{EvaluatePairsFromXCMSSet} with respect to interesting m/z-pairs.
}
\keyword{data}
