% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xg-data.R
\docType{data}
\name{xg}
\alias{xg}
\title{Peak table}
\description{
This data contains a grouped xcms peak table built on the files in \link{raw}.
}
\author{
Jan Lisec \email{jan.lisec@bam.de}
}
\keyword{data}
