% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateCandXLSX.R
\name{GenerateCandXLSX}
\alias{GenerateCandXLSX}
\title{GenerateCandXLSX.}
\usage{
GenerateCandXLSX(res_list = NULL, xlsx_file = NULL, rejected = FALSE)
}
\arguments{
\item{res_list}{A list of result objects (each testing an individual mz pair).}

\item{xlsx_file}{File name.}

\item{rejected}{Logical. Return rejected if TRUE.}
}
\value{
Candidate table as data.frame.
}
\description{
\code{GenerateCandXLSX} will produce a XLSX of a list containing test results objects.
}
\details{
Not yet.
}
\examples{
# load evaluation result of example data
data(res_list)
# generate table within R (use xlsx_file to write to file)
str(GenerateCandXLSX(res_list))
GenerateCandXLSX(res_list)[, 1:5]

}
