% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mz_shift_corrector-data.R
\docType{data}
\name{mz_shift_corrector}
\alias{mz_shift_corrector}
\title{Predefined mass search windows to be used internally.}
\description{
This is a list defining windows for high res APCI or ESI instrumentation..
}
\author{
Jan Lisec \email{jan.lisec@bam.de}
}
\keyword{data}
