% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res_list-data.R
\docType{data}
\name{res_list}
\alias{res_list}
\title{The main results object of a non-targeted search for tracer incorporation.}
\description{
This is a list containing the evaluations results established based on processing example data with \link{EvaluateCandidateListAgainstRawData}.
}
\author{
Jan Lisec \email{jan.lisec@bam.de}
}
\keyword{data}
