% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sam-data.R
\docType{data}
\name{sam}
\alias{sam}
\title{Sample table}
\description{
This data frame contains the sample definition of 18 samples from a larger experiment.
}
\author{
Jan Lisec \email{jan.lisec@bam.de}
}
\keyword{data}
