% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcms_cand-data.R
\docType{data}
\name{xcms_cand}
\alias{xcms_cand}
\title{Dataframe with putative candidates}
\description{
This data frame contains the analysis result of an xcmsSet which can not be provided via CRAN anymore using \link{EvaluatePairsFromXCMSSet} with respect to interesting m/z-pairs.
}
\author{
Jan Lisec \email{jan.lisec@bam.de}
}
\keyword{data}
